+++ src/Microsoft.OpenApi/Reader/V2/OpenApiOperationDeserializer.cs
@@ -23,8 +23,18 @@ namespace Microsoft.OpenApi.Reader.V2
             new()
             {
                 {
-                    "tags", (o, n, doc) => { 
-                        if (n.CreateSimpleList((valueNode, doc) => LoadTagByReference(valueNode.GetScalarValue(), doc), doc) is {Count: > 0} tags)
+                    "tags", (o, n, doc) => {
+                        if (n.CreateSimpleList(
+                            (valueNode, doc) =>
+                            {
+                                var val = valueNode.GetScalarValue();
+                                if (string.IsNullOrEmpty(val))
+                                    return null;   // Avoid exception on empty tag, we'll remove these from the list further on
+                                return LoadTagByReference(val , doc);
+                                },
+                            doc)
+                        // Filter out empty tags instead of excepting on them
+                        .OfType<OpenApiTagReference>().ToList() is {Count: > 0} tags)
                         {
                             o.Tags = new HashSet<OpenApiTagReference>(tags, OpenApiTagComparer.Instance);
                         }